% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faculty.R
\name{facto}
\alias{facto}
\title{gives factorial of n
based on a simple fortran function \code{faculty} calculates the factorial of n.}
\usage{
facto(n, type = c("Fortran", "R"))
}
\description{
gives factorial of n
based on a simple fortran function \code{faculty} calculates the factorial of n.
}
