% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{squares number(s)}
\usage{
square(x, type = c("Fortran", "R"))
}
\arguments{
\item{x}{single number or even a vector containing numerical values}

\item{type}{character value: language which shall be used for calculating: 'Fortran' or 'R'}
}
\description{
based on a fortran subroutine or calculated in R, depending in type
}
\examples{
require(microbenchmark)
microbenchmark(square(1:1000, type = "Fortran"), square(1:1000, type = "R"))
}
